refactor(core): move service integrations to owner crate
