+++ src/crates/core/src/service/mcp/config/location.rs
@@ -1,10 +1,1 @@
-use serde::{Deserialize, Serialize};
-
-/// Configuration location.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "kebab-case")]
-pub enum ConfigLocation {
-    BuiltIn, // Built-in configuration
-    User,    // User-level configuration
-    Project, // Project-level configuration
-}
+pub use bitfun_services_integrations::mcp::config::ConfigLocation;
