+++ src/crates/core/src/service/mcp/server/mod.rs
@@ -8,8 +8,9 @@ mod manager;
 mod process;
 mod registry;
 
+pub use bitfun_services_integrations::mcp::server::{MCPServerStatus, MCPServerType};
 pub use config::{MCPServerConfig, MCPServerOAuthConfig, MCPServerTransport, MCPServerXaaConfig};
 pub use connection::{MCPConnection, MCPConnectionPool};
 pub use manager::MCPServerManager;
-pub use process::{MCPServerProcess, MCPServerStatus, MCPServerType};
+pub use process::MCPServerProcess;
 pub use registry::MCPServerRegistry;
