+++ src/crates/core/src/service/mcp/tool_info.rs
@@ -1,8 +1,1 @@
-use serde::{Deserialize, Serialize};
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct McpToolInfo {
-    pub server_id: String,
-    pub server_name: String,
-    pub tool_name: String,
-}
+pub use bitfun_services_integrations::mcp::McpToolInfo;
