+++ src/crates/services-integrations/src/announcement/mod.rs
@@ -1,0 +1,5 @@
+//! Announcement service data contracts.
+
+mod types;
+
+pub use types::*;
