+++ src/crates/services-integrations/src/mcp/config/mod.rs
@@ -1,0 +1,5 @@
+//! MCP configuration data contracts.
+
+mod location;
+
+pub use location::ConfigLocation;
