+++ src/crates/services-integrations/src/mcp/protocol/mod.rs
@@ -1,0 +1,5 @@
+//! MCP protocol data contracts.
+
+pub mod types;
+
+pub use types::*;
