+++ src/crates/services-integrations/src/remote_ssh/mod.rs
@@ -1,0 +1,8 @@
+//! Remote SSH service contracts.
+//!
+//! `bitfun-core::service::remote_ssh` remains as the compatibility facade for
+//! the legacy public path.
+
+pub mod types;
+
+pub use types::*;
