+++ src/crates/services-integrations/tests/mcp_contracts.rs
@@ -1,0 +1,145 @@
+#![cfg(feature = "mcp")]
+
+use bitfun_services_integrations::mcp::config::ConfigLocation;
+use bitfun_services_integrations::mcp::protocol::{
+    MCPCapability, MCPError, MCPPromptMessageContent, MCPPromptMessageContentBlock, MCPRequest,
+    default_protocol_version,
+};
+use bitfun_services_integrations::mcp::server::{MCPServerStatus, MCPServerType};
+use bitfun_services_integrations::mcp::{
+    MCP_TOOL_DELIMITER, MCP_TOOL_PREFIX, McpToolInfo, build_mcp_tool_name, normalize_name_for_mcp,
+};
+
+#[test]
+fn mcp_tool_name_contract_matches_existing_wire_format() {
+    assert_eq!(MCP_TOOL_PREFIX, "mcp__");
+    assert_eq!(MCP_TOOL_DELIMITER, "__");
+    assert_eq!(
+        normalize_name_for_mcp("Acme Search / Primary"),
+        "Acme_Search___Primary"
+    );
+    assert_eq!(
+        build_mcp_tool_name("Claude Code", "search repos"),
+        "mcp__Claude_Code__search_repos"
+    );
+}
+
+#[test]
+fn mcp_tool_info_preserves_json_shape() {
+    let info = McpToolInfo {
+        server_id: "server-1".to_string(),
+        server_name: "Docs".to_string(),
+        tool_name: "search".to_string(),
+    };
+
+    assert_eq!(
+        serde_json::to_value(info).unwrap(),
+        serde_json::json!({
+            "server_id": "server-1",
+            "server_name": "Docs",
+            "tool_name": "search"
+        })
+    );
+}
+
+#[test]
+fn mcp_protocol_capability_contract_matches_existing_default() {
+    assert_eq!(default_protocol_version(), "2025-11-25");
+    assert_eq!(
+        serde_json::to_value(MCPCapability::default()).unwrap(),
+        serde_json::json!({
+            "resources": {
+                "subscribe": false,
+                "listChanged": false
+            },
+            "prompts": {
+                "listChanged": false
+            },
+            "tools": {
+                "listChanged": false
+            }
+        })
+    );
+}
+
+#[test]
+fn mcp_protocol_jsonrpc_helpers_preserve_wire_shape() {
+    let request = MCPRequest::new(
+        serde_json::json!(7),
+        "tools/list".to_string(),
+        Some(serde_json::json!({ "cursor": "next" })),
+    );
+
+    assert_eq!(
+        serde_json::to_value(request).unwrap(),
+        serde_json::json!({
+            "jsonrpc": "2.0",
+            "id": 7,
+            "method": "tools/list",
+            "params": {
+                "cursor": "next"
+            }
+        })
+    );
+
+    assert_eq!(
+        serde_json::to_value(MCPError::method_not_found("tools/call")).unwrap(),
+        serde_json::json!({
+            "code": -32601,
+            "message": "Method not found: tools/call"
+        })
+    );
+}
+
+#[test]
+fn mcp_protocol_prompt_content_helpers_preserve_legacy_text_behavior() {
+    let mut content = MCPPromptMessageContent::Plain("Review {{target}}".to_string());
+    content.substitute_placeholders(&std::collections::HashMap::from([(
+        "target".to_string(),
+        "src/main.rs".to_string(),
+    )]));
+
+    assert_eq!(content.text_or_placeholder(), "Review src/main.rs");
+
+    let image = MCPPromptMessageContent::Block(Box::new(MCPPromptMessageContentBlock::Image {
+        data: "base64".to_string(),
+        mime_type: "image/png".to_string(),
+    }));
+    assert_eq!(image.text_or_placeholder(), "[Image: image/png]");
+}
+
+#[test]
+fn mcp_config_location_preserves_kebab_case_wire_contract() {
+    assert_eq!(
+        serde_json::to_value(ConfigLocation::BuiltIn).unwrap(),
+        serde_json::json!("built-in")
+    );
+    assert_eq!(
+        serde_json::from_value::<ConfigLocation>(serde_json::json!("user")).unwrap(),
+        ConfigLocation::User
+    );
+    assert_eq!(
+        serde_json::from_value::<ConfigLocation>(serde_json::json!("project")).unwrap(),
+        ConfigLocation::Project
+    );
+}
+
+#[test]
+fn mcp_server_type_and_status_preserve_lowercase_wire_contract() {
+    assert_eq!(
+        serde_json::to_value(MCPServerType::Local).unwrap(),
+        serde_json::json!("local")
+    );
+    assert_eq!(
+        serde_json::from_value::<MCPServerType>(serde_json::json!("remote")).unwrap(),
+        MCPServerType::Remote
+    );
+    assert_eq!(
+        serde_json::to_value(MCPServerStatus::NeedsAuth).unwrap(),
+        serde_json::json!("needsauth")
+    );
+    assert_eq!(
+        serde_json::from_value::<MCPServerStatus>(serde_json::json!("reconnecting")).unwrap(),
+        MCPServerStatus::Reconnecting
+    );
+}
